feat: Migrate from CRA to Vite
Simple change from react-scripts to Vite since create-react-app has no
active mantainers and has stopped releasing new versions. Only a few
changes were needed to make sure Vite worked exactly as react-script
did previously. Hopefully this makes the move to typescript and
server-side rendering for the landing page much easier as Vite has
native support for typescript and plugins for server side rendering.

I've opted to migrate to Vite instead of using a full-blown framework
such as Next.JS because changing to a framework would require a lot of
changes to the codebase, and currently it is already very neat and tidy,
adding the structure from Next.JS would just complicate things without a
lot of advantages for this SPA.